requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
groq>=0.3.0  # Add Groq package
python-dotenv>=1.0.0
streamlit>=1.25.0
duckduckgo-search>=3.8.5  # Keep for web search fallback